[Update] 네비게이션바 디자인 변경
