+++ lib/view/home_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../components/noreuse/homeTopBar.dart';
+import '../components/noreuse/home_top_bar.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({Key? key}) : super(key: key);
